Refactor AP3 `Window` and Params (#808)
