+++ core/txpool/blobpool/blobpool.go
@@ -39,7 +39,6 @@ import (
 	"sync"
 	"time"
 
-	"github.com/ava-labs/subnet-evm/consensus/dummy"
 	"github.com/ava-labs/subnet-evm/consensus/misc/eip4844"
 	"github.com/ava-labs/subnet-evm/core"
 	"github.com/ava-labs/subnet-evm/core/state"
@@ -47,6 +46,7 @@ import (
 	"github.com/ava-labs/subnet-evm/core/types"
 	"github.com/ava-labs/subnet-evm/metrics"
 	"github.com/ava-labs/subnet-evm/params"
+	"github.com/ava-labs/subnet-evm/plugin/evm/header"
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/event"
 	"github.com/ethereum/go-ethereum/log"
@@ -415,7 +415,7 @@ func (p *BlobPool) Init(gasTip uint64, head *types.Header, reserve txpool.Addres
 		p.Close()
 		return err
 	}
-	baseFee, err := dummy.EstimateNextBaseFee(
+	baseFee, err := header.EstimateNextBaseFee(
 		p.chain.Config(),
 		feeConfig,
 		p.head,
@@ -851,7 +851,7 @@ func (p *BlobPool) Reset(oldHead, newHead *types.Header) {
 		log.Error("Failed to get fee config to reset blobpool fees", "err", err)
 		return
 	}
-	baseFeeBig, err := dummy.EstimateNextBaseFee(
+	baseFeeBig, err := header.EstimateNextBaseFee(
 		p.chain.Config(),
 		feeConfig,
 		p.head,
