+++ plugin/evm/header/block_gas_cost.go
@@ -4,11 +4,18 @@
 package header
 
 import (
+	"errors"
 	"math/big"
 
 	"github.com/ava-labs/subnet-evm/commontype"
 	"github.com/ava-labs/subnet-evm/core/types"
+	"github.com/ava-labs/subnet-evm/params"
 	"github.com/ava-labs/subnet-evm/plugin/evm/blockgascost"
+)
+
+var (
+	errBaseFeeNil      = errors.New("base fee is nil")
+	errBlockGasCostNil = errors.New("block gas cost is nil")
 )
 
 // BlockGasCost calculates the required block gas cost based on the parent
@@ -58,4 +65,32 @@ func BlockGasCostWithStep(
 		step,
 		timeElapsed,
 	)
+}
+
+// MinRequiredTip is the estimated minimum tip a transaction would have
+// needed to pay to be included in a given block (assuming it paid a tip
+// proportional to its gas usage). In reality, there is no minimum tip that
+// is enforced by the consensus engine and high tip paying transactions can
+// subsidize the inclusion of low tip paying transactions. The only
+// correctness check performed is that the sum of all tips is >= the
+// required block fee.
+//
+// This function will return nil for all return values prior to Subnet EVM.
+func MinRequiredTip(config *params.ChainConfig, header *types.Header) (*big.Int, error) {
+	if !config.IsSubnetEVM(header.Time) {
+		return nil, nil
+	}
+	if header.BaseFee == nil {
+		return nil, errBaseFeeNil
+	}
+	if header.BlockGasCost == nil {
+		return nil, errBlockGasCostNil
+	}
+
+	// minTip = requiredBlockFee/blockGasUsage
+	requiredBlockFee := new(big.Int).Mul(
+		header.BlockGasCost,
+		header.BaseFee,
+	)
+	return new(big.Int).Div(requiredBlockFee, new(big.Int).SetUint64(header.GasUsed)), nil
 }
