+++ plugin/evm/header/extra_test.go
@@ -34,7 +34,7 @@ func TestVerifyExtra(t *testing.T) {
 			rules: params.AvalancheRules{
 				IsSubnetEVM: true,
 			},
-			extra:    make([]byte, params.DynamicFeeExtraDataSize),
+			extra:    make([]byte, FeeWindowSize),
 			expected: nil,
 		},
 		{
@@ -42,7 +42,7 @@ func TestVerifyExtra(t *testing.T) {
 			rules: params.AvalancheRules{
 				IsSubnetEVM: true,
 			},
-			extra:    make([]byte, params.DynamicFeeExtraDataSize-1),
+			extra:    make([]byte, FeeWindowSize-1),
 			expected: errInvalidExtraLength,
 		},
 		{
@@ -50,7 +50,7 @@ func TestVerifyExtra(t *testing.T) {
 			rules: params.AvalancheRules{
 				IsSubnetEVM: true,
 			},
-			extra:    make([]byte, params.DynamicFeeExtraDataSize+1),
+			extra:    make([]byte, FeeWindowSize+1),
 			expected: errInvalidExtraLength,
 		},
 		{
@@ -58,7 +58,7 @@ func TestVerifyExtra(t *testing.T) {
 			rules: params.AvalancheRules{
 				IsDurango: true,
 			},
-			extra:    make([]byte, params.DynamicFeeExtraDataSize),
+			extra:    make([]byte, FeeWindowSize),
 			expected: nil,
 		},
 		{
@@ -66,7 +66,7 @@ func TestVerifyExtra(t *testing.T) {
 			rules: params.AvalancheRules{
 				IsDurango: true,
 			},
-			extra:    make([]byte, params.DynamicFeeExtraDataSize+1),
+			extra:    make([]byte, FeeWindowSize+1),
 			expected: nil,
 		},
 		{
@@ -74,7 +74,7 @@ func TestVerifyExtra(t *testing.T) {
 			rules: params.AvalancheRules{
 				IsDurango: true,
 			},
-			extra:    make([]byte, params.DynamicFeeExtraDataSize-1),
+			extra:    make([]byte, FeeWindowSize-1),
 			expected: errInvalidExtraLength,
 		},
 	}
