+++ plugin/evm/upgrade/legacy/params.go
@@ -1,0 +1,5 @@
+package legacy
+
+const (
+	BaseFee int64 = 225_000_000_000
+)
