+++ predicate/predicate_bytes.go
@@ -6,7 +6,7 @@ package predicate
 import (
 	"fmt"
 
-	"github.com/ava-labs/subnet-evm/params"
+	"github.com/ava-labs/subnet-evm/plugin/evm/header"
 	"github.com/ethereum/go-ethereum/common"
 )
 
@@ -57,8 +57,8 @@ func GetPredicateResultBytes(extraData []byte) ([]byte, bool) {
 	// Prior to Durango, the VM enforces the extra data is smaller than or equal to this size.
 	// After Durango, the VM pre-verifies the extra data past the dynamic fee rollup window is
 	// valid.
-	if len(extraData) <= params.DynamicFeeExtraDataSize {
+	if len(extraData) <= header.FeeWindowSize {
 		return nil, false
 	}
-	return extraData[params.DynamicFeeExtraDataSize:], true
+	return extraData[header.FeeWindowSize:], true
 }
